Move IConfigService to AfipHelper
